Narrow to the draft-mint gate release
The text-restore behavior moves to the unsent-draft work, where a failed
message stays on screen with its attachments instead of being rebuilt from
the composer. That resolves the retry-identity and attachment questions by
construction rather than by guarding a rebuild.
